Wait for database to be ready before migrating
Previous implementation was just waiting for the postgres port to
be ready, but that is not enough, because the port is available before
the database can actually accept querys.

This version of the secript uses pg_isready to ensure postgres is ready.

Wait for it is not required any more, but I am keeping it because it can
be handy for other checks.